Bugfix: the end of a reorged chain is invalid when connect fails
Introduced in bedb62a453a537000b5999edb593c9131b1439c7

When an invalid block is found during a reorg, we know the last of the blocks in
the was-to-be-connected chain is invalid, but not necessarily the first. As
vpIndexToConnect is ordered in decreasing height, the end of the reorg is the
front of the vector, and not the back.

This only affected the warning system.